Create README.txt
This is the RSLK Car Project for ECE 3 at UCLA. This project was done by myself and took roughly 6 weeks to build, code, and test my robot line follower. The objective for this project was to code the robotic car to follow a black line, which was printed out as an obstacle course. Given source codes include the encoder and QTR sensor files. The coding was done in C/C++ using the Arduino IDE.